refactor: migrate all duration tracking from seconds to milliseconds across domain and application layers

+++ crates/adapters-acoustid/src/response.rs
@@ -19,5 +19,5 @@ pub struct AcoustIdResult {
 pub struct AcoustIdRecording {
     pub id: String, // MusicBrainz Recording ID
     #[serde(default)]
-    pub duration: Option<u32>, // Used for best-match selection (B3 fix)
+    pub duration: Option<f64>,
 }
